house_back/requirements.txt
Django==4.2.13
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.4.0
Pillow==10.4.0
python-decouple==3.8
scikit-learn==1.5.2
numpy>=2.0.0
django-filter==24.3
cloudinary==1.41.0
django-cloudinary-storage==0.3.0
channels==4.1.0
daphne==4.1.2
whitenoise==6.7.0
groq==0.9.0
